Move the DB tables schema definition into own file
